refac console.warn mocking
