+++ test/configCases/sharing/consume-multiple-versions/index.js
@@ -1,31 +1,6 @@
-let warnings = [];
-let oldWarn;
+const expectWarningFactory = require('../../../helpers/expectWarningFactory');
 
-beforeEach(done => {
-	oldWarn = console.warn;
-	console.warn = m => warnings.push(m);
-	done();
-});
-
-afterEach(done => {
-	expectWarning();
-	console.warn = oldWarn;
-	done();
-});
-
-const expectWarning = regexp => {
-	if (!regexp) {
-		expect(warnings).toEqual([]);
-	} else {
-		expect(warnings).toEqual(
-			expect.objectContaining({
-				0: expect.stringMatching(regexp),
-				length: 1
-			})
-		);
-	}
-	warnings.length = 0;
-};
+const expectWarning = expectWarningFactory();
 
 it("should be able to consume different shared module version depending on context", async () => {
 	__webpack_share_scopes__["default"] = {
